Create new client index.js
